refator: 총 상금 변수 int 에서 long으로 변경

+++ src/main/java/lotto/domain/Statistics.java
@@ -30,7 +30,7 @@ public class Statistics {
     }
 
     public long getEarnMoney() {
-        int earnMoney = 0;
+        long earnMoney = 0;
 
         for (Rank rank : Rank.values()) {
             earnMoney += lottoResult.getOrDefault(rank, GAME_COUNT_ZERO) * rank.getWinningMoney();
